+++ wordclock/content/wificfg/hwcfg.html
@@ -52,6 +52,7 @@
 "<button name=\"cl_command\" type=\"submit\" value=\"Save\">Save</button>"
 "<button name=\"cl_command\" type=\"submit\" value=\"Reboot\">Reboot</button>"
 "<button name=\"cl_command\" type=\"submit\" value=\"OtaUpdate\">sw update</button>"
+"<button name=\"cl_command\" type=\"submit\" value=\"FactoryReset\">Factory Reset</button>"
 "</center>"
 "</fieldset>"
 "</form>"
