+++ wordclock/content/wificfg/sta.html
@@ -28,6 +28,7 @@
 "\"></dd>"
 "</dl>"
 "<button name=\"sta_command\" type=\"submit\" value=\"Save\">Save</button>"
+"<button name=\"sta_command\" type=\"submit\" value=\"Refresh\">Refresh</button>"
 "</fieldset>"
 "</form>"
 "</body></html>"
